requirements.txt
crypto==1.4.1
cryptography==39.0.2
opcua==0.98.13
Pillow==9.5.0
pyModbusTCP==0.2.0
python-dateutil==2.8.2

